Refactor the signing code, add coments to the function that generates the initialisation hash
