+++ examples/4337-passkeys/src/logic/passkeys.ts
@@ -1,3 +1,6 @@
+import { ethers } from "ethers";
+import { hexStringToUint8Array } from "../utils.ts";
+
 type PasskeyCredential = {
   id: 'string'
   rawId: ArrayBuffer
@@ -81,12 +84,75 @@ async function createPasskey(): Promise<PasskeyCredentialWithPubkeyCoordinates> 
   return passkeyWithCoordinates
 }
 
+/**
+ * Extracts the signature into R and S values from the authenticator response.
+ *
+ * See:
+ * - <https://datatracker.ietf.org/doc/html/rfc3279#section-2.2.3>
+ * - <https://en.wikipedia.org/wiki/X.690#BER_encoding>
+ */
+function extractSignature(response: AuthenticatorAssertionResponse): [bigint, bigint] {
+  const check = (x: boolean) => {
+    if (!x) {
+      throw new Error('invalid signature encoding')
+    }
+  }
+
+  // Decode the DER signature. Note that we assume that all lengths fit into 8-bit integers,
+  // which is true for the kinds of signatures we are decoding but generally false. I.e. this
+  // code should not be used in any serious application.
+  const view = new DataView(response.signature)
+
+  // check that the sequence header is valid
+  check(view.getUint8(0) === 0x30)
+  check(view.getUint8(1) === view.byteLength - 2)
+
+  // read r and s
+  const readInt = (offset: number) => {
+    check(view.getUint8(offset) === 0x02)
+    const len = view.getUint8(offset + 1)
+    const start = offset + 2
+    const end = start + len
+    const n = BigInt(ethers.hexlify(new Uint8Array(view.buffer.slice(start, end))))
+    check(n < ethers.MaxUint256)
+    return [n, end] as const
+  }
+  const [r, sOffset] = readInt(2)
+  const [s] = readInt(sOffset)
+
+  return [r, s]
+}
+
+type Assertion = {
+  response: AuthenticatorAssertionResponse
+}
+
+
 export type PasskeyLocalStorageFormat = {
   rawId: string
   pubkeyCoordinates: {
     x: string
     y: string
   }
+}
+
+/**
+ * Compute the additional client data JSON fields. This is the fields other than `type` and
+ * `challenge` (including `origin` and any other additional client data fields that may be
+ * added by the authenticator).
+ *
+ * See <https://w3c.github.io/webauthn/#clientdatajson-serialization>
+ */
+function extractClientDataFields(response: AuthenticatorAssertionResponse): string {
+  const clientDataJSON = new TextDecoder('utf-8').decode(response.clientDataJSON)
+  const match = clientDataJSON.match(/^\{"type":"webauthn.get","challenge":"[A-Za-z0-9\-_]{43}",(.*)\}$/)
+
+  if (!match) {
+    throw new Error('challenge not found in client data JSON')
+  }
+
+  const [, fields] = match
+  return ethers.hexlify(ethers.toUtf8Bytes(fields))
 }
 
 /**
@@ -103,6 +169,40 @@ function toLocalStorageFormat(passkey: PasskeyCredentialWithPubkeyCoordinates): 
 }
 
 /**
+ * Signs data with a passkey.
+ *
+ * This function uses the WebAuthn API to sign data with a passkey. The passkey is identified by its ID.
+ * The function throws an error if the signing operation fails.
+ *
+ * @param passkeyId - The ID of the passkey to use for signing.
+ * @param data - The data to sign.
+ * @returns A promise that resolves to the signed data. The signed data is encoded using ethers.js's default ABI coder.
+ * @throws Throws an error if the signing operation fails.
+ */
+async function signWithPasskey(passkeyId: string, data: string): Promise<string> {
+  const assertion = (await navigator.credentials.get({
+    publicKey: {
+      challenge: ethers.getBytes(data),
+      allowCredentials: [{ type: 'public-key', id: hexStringToUint8Array(passkeyId) }],
+      userVerification: 'required',
+    },
+  })) as Assertion | null
+
+  if (!assertion) {
+    throw new Error('Failed to sign user operation')
+  }
+
+  return ethers.AbiCoder.defaultAbiCoder().encode(
+    ['bytes', 'bytes', 'uint256[2]'],
+    [
+      new Uint8Array(assertion.response.authenticatorData),
+      extractClientDataFields(assertion.response),
+      extractSignature(assertion.response),
+    ],
+  )
+}
+
+/**
  * Checks if the provided value is in the format of a Local Storage Passkey.
  * @param x The value to check.
  * @returns A boolean indicating whether the value is in the format of a Local Storage Passkey.
@@ -111,4 +211,5 @@ function isLocalStoragePasskey(x: unknown): x is PasskeyLocalStorageFormat {
   return typeof x === 'object' && x !== null && 'rawId' in x && 'pubkeyCoordinates' in x
 }
 
-export { createPasskey, toLocalStorageFormat, isLocalStoragePasskey }
+export type { Assertion }
+export { createPasskey, toLocalStorageFormat, isLocalStoragePasskey, extractSignature, extractClientDataFields, signWithPasskey }
